Laravel policy successfully initiate!

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,8 @@ use App\Models\Post;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use Illuminate\Support\Facades\Auth;
+// use App\Http\Resources\PostResource;
+
 
 class PostController extends Controller
 {
@@ -14,14 +16,14 @@ class PostController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Post::class);
         $posts = Post::where('user_id', auth::user()->id)->get();
         return view('resources.post.index', ['posts' => $posts]);
-        
     }
 
     /**
      * Show the form for creating a new resource.
-     */
+     */ 
     public function create()
     {
         return view('resources.post.create');
@@ -36,7 +38,7 @@ class PostController extends Controller
             'user_id' => auth::user()->id,
             'subject' => $request->subject,
             'post' => $request->post,
-            'status' => (is_null($request->status) ? 0 : 1)
+            'status' => ($request->status == "on" ? 1 : 0)
         ]);
         return redirect()->route('post.index')->with('message', 'Post Successfully Saved!');
     }
@@ -46,6 +48,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('view', $post);
         return view('resources.post.show', ['post' => $post]);
     }
 
@@ -54,6 +57,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('view', $post);
         return view('resources.post.edit', ['post' => $post]);
     }
 
@@ -62,11 +66,12 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+        $this->authorize('view', $post);
         $post->update([
             'user_id' => auth::user()->id,
             'subject' => $request->subject,
             'post' => $request->post,
-            'status' => (is_null($request->status) ? 0 : 1)
+            'status' => ($request->status == "on" ? 1 : 0)
         ]);
         return redirect()->route('post.index')->with('message', 'Post Successfully Saved!');
     }
@@ -76,6 +81,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('view', $post);
         $post->delete();
         return redirect()->route('post.index')->with('message', 'Post Successfully Deleted!');
     }
